Move all private functions to webGL_curves.ts
